Move declaration conversion into its own module
Amp-Thread-ID: https://ampcode.com/threads/T-01a03a21-1ee3-7618-b6fd-5931af44e545
